+ Get obstacle range from setting

+++ RightTap/Assets/Scripts/Obstacle.cs
@@ -20,11 +20,11 @@ public class Obstacle : MonoBehaviour
         this.shouldMove = false;
     }
 
-    public void SetParams(int speed, int minRange, int maxRange)
+    public void SetParams(int speed, float minRange, float maxRange)
     {
         this.direction = new Vector3(0.0f, -speed / 100.0f);
-        this.minRange = minRange;
-        this.maxRange = maxRange;
+        this.minRange = (int)minRange;
+        this.maxRange = (int)maxRange;
     }
 
     private void RefreshRange()
